Added project with YOLOv5

+++ src/YOLO/IYOLO.cs
@@ -4,6 +4,6 @@ namespace YOLO
 {
     public interface IYOLO<in TInputImage> where TInputImage : class
     {
-        IEnumerable<DetectedObject> Forward(TInputImage image, double confidenceThreshold, double nmsThreshold);
+        IEnumerable<DetectedObject> Forward(TInputImage image, float confidenceThreshold, float nmsThreshold);
     }
 }
